2024 0509 1454 재후

+++ src/main/java/com/sist/nbgb/controller/UserController.java
@@ -4,6 +4,7 @@ import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.UnsupportedEncodingException;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -34,7 +35,6 @@ import com.sist.nbgb.dto.OfflinePaymentCancelDto;
 import com.sist.nbgb.dto.OnlinePaymentApproveDto;
 import com.sist.nbgb.dto.OnlinePaymentCancelDto;
 import com.sist.nbgb.dto.PhoneChangeDto;
-import com.sist.nbgb.dto.UserFileDto;
 import com.sist.nbgb.dto.UserIdCheckDto;
 import com.sist.nbgb.dto.UserReviewDto;
 import com.sist.nbgb.service.EmailService;
@@ -63,15 +63,18 @@ public class UserController
     	if(id != null)
     	{
     		UserInfoDto userInfoDto = userService.findByUserId(id);
-    		
-    		UserFileDto userFileDto = userService.findByUserIdFile(id);
-    		
         	model.addAttribute("userInfoDto", userInfoDto);
         	
-        	if(userFileDto != null)
-        	{
-        		model.addAttribute("userFileDto", userFileDto);
-        	}
+    		String path = "C:\\project\\sts4\\SFPN\\src\\main\\resources\\static\\images\\user";
+    	    String filename = userInfoDto.getUserId() + ".png"; // 기본 파일명
+    	    String filepath = path + "/" + filename;
+    		
+            File file = new File(filepath);
+            
+    		if(file.exists())
+    		{
+    			model.addAttribute("userFileDto", userInfoDto.getUserId());
+    		}
         	else
         	{
         		model.addAttribute("userFileDto", "userFileDtoIsNull");
@@ -88,7 +91,24 @@ public class UserController
     	{
     		UserInfoDto userInfoDto = userService.findByUserId(id);
         	model.addAttribute("userInfoDto", userInfoDto);
+        	
+    		String path = "C:\\project\\sts4\\SFPN\\src\\main\\resources\\static\\images\\user";
+    	    String filename = userInfoDto.getUserId() + ".png"; // 기본 파일명
+    	    String filepath = path + "/" + filename;
+    		
+            File file = new File(filepath);
+            
+    		if(file.exists())
+    		{
+    			model.addAttribute("userFileDto", userInfoDto.getUserId());
+    		}
+        	else
+        	{
+        		model.addAttribute("userFileDto", "userFileDtoIsNull");
+        	}
     	}
+    	
+    	
     	
     	return "mypage/mypageModify";
     }
@@ -158,25 +178,17 @@ public class UserController
 	@ResponseBody
 	public ResponseEntity<Object> mypageUserInfoFile(@RequestPart(value="userId") @Valid UserIdCheckDto userIdCheckDto, @RequestPart(value="userImageFile") MultipartFile userImageFile) throws Exception
 	{
-    	UserFileDto userFileDto = userService.findByUserIdFile(userIdCheckDto.getUserId()); 
-    	
-    	if(userFileDto != null)
-    	{
-    		
-    	}
-    	else
-    	{
-    	
-    	}
-    	
-    	
-	    String path = "C:\\project\\sts4\\SFPN\\src\\main\\resources\\static\\images\\user";
+    	String path = "C:\\project\\sts4\\SFPN\\src\\main\\resources\\static\\images\\user";
 	    String filename = userIdCheckDto.getUserId() + ".png"; // 기본 파일명
 	    String filepath = path + "/" + filename;
 		
         File file = new File(filepath);
-        file.delete();
         
+		if(file.exists())
+		{
+			file.delete();
+    	}
+
         try 
         {
             BufferedOutputStream bufferedOutputStream = new BufferedOutputStream(new FileOutputStream(file));
@@ -189,8 +201,21 @@ public class UserController
         {
             throw new RuntimeException("오류가 발생했습니다.");
         }
-	}
+    }
     
+    @PostMapping("/mypage/info/userfiledelete")
+    @ResponseBody
+    public ResponseEntity<Object> mypageUserInfoFileDelete(@RequestBody UserIdCheckDto userIdCheckDto)
+    {
+    	String path = "C:\\project\\sts4\\SFPN\\src\\main\\resources\\static\\images\\user";
+	    String filename = userIdCheckDto.getUserId() + ".png"; // 기본 파일명
+	    String filepath = path + "/" + filename;
+		
+        File file = new File(filepath);
+        file.delete();
+    	
+    	return ResponseEntity.ok(200);
+    }
     
     
     @GetMapping("/mypage/{id}/signout")
@@ -278,14 +303,16 @@ public class UserController
     		{
     			model.addAttribute("offlinePaymentCancelDto", offlinePaymentCancelDto);
     		}
+    		
+    		model.addAttribute("localDateTime", LocalDateTime.now());
     	}
     	
     	return "mypage/mypagePayment";
     }
     
     
-    @GetMapping("/mypage/{id}/lecture")
-    public String mypageUserLecture(Model model, @PathVariable String id)
+    @GetMapping("/mypage/{id}/onlinelecturelist")
+    public String mypageUserOnlineLecture(Model model, @PathVariable String id)
     {
     	if(id != null)
     	{
@@ -304,10 +331,51 @@ public class UserController
         	model.addAttribute("userReviewDto", userReviewDto);
     	}
     	
-    	return "mypage/mypageLecture";
+    	return "mypage/mypageOnlineLectureList";
     }
     
+    @GetMapping("/mypage/{id}/offlinelecturelist")
+    public String mypageUserOfflineLecture(Model model, @PathVariable String id)
+    {
+    	if(id != null)
+    	{
+    		List<UserReviewDto> userReviewDto = userService.userReviewFindAll(id)
+    				.stream()
+    				.map(UserReviewDto::new)
+    				.collect(Collectors.toList());
+    		
+//    		List<OnlineClassTitleDto> onlineClassTitleDto = onlineClassService.mypageTitle()
+//    				.stream().map(OnlineClassTitleDto::new).collect(Collectors.toList());
+    				
+    		
+    	//	List<OfflineResponse> offlineClassListDto = offlineService.
+    		
+    		
+        	model.addAttribute("userReviewDto", userReviewDto);
+    	}
+    	
+    	return "mypage/mypageOfflineLectureList";
+    }
     
+    @GetMapping("/mypage/{id}/onlinelikelist")
+    public String mypageUserOnlineLike(Model model, @PathVariable String id)
+    {
+    	if(id != null)
+    	{
+    	}
+    	
+    	return "mypage/mypageOnlineLikeList";
+    }
+    
+    @GetMapping("/mypage/{id}/offlinelikelist")
+    public String mypageUserOfflineLike(Model model, @PathVariable String id)
+    {
+    	if(id != null)
+    	{
+    	}
+    	
+    	return "mypage/mypageOfflineLikeList";
+    }
     
     @GetMapping("/mypage/{id}/review")
     public String mypageUserReview(Model model, @PathVariable String id)
